Move 'RemoteFeedItesm' to its own file
